Update oc2pf.md
According to Section 2.3.4.1 OpenC2 Producers that send the delete: pf:rule_number command  MUST NOT include other Command Arguments. The table specifies that the start_time argument can be used. I prefer to go with what Section 2.3.4.1 defines - as the right way of doing things. An option to have a property that specified the timestamp to execute this command would be part of the orchestrator/orchestration language and to the extent of my knowledge, this is not a common PF capability.